% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{.RIGHT}
\alias{.RIGHT}
\title{RIGHT}
\usage{
.RIGHT(string, n)
}
\arguments{
\item{string}{A string to extract text from}

\item{n}{the number of characters to extract}
}
\description{
Extracts the right \code{n} characters from a character string
}
