% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blscraper_funcs.R
\name{get_bls_data}
\alias{get_bls_data}
\title{get_bls_data}
\usage{
get_bls_data(series_id, start_year = NULL, end_year = NULL, api_key = NULL)
}
\arguments{
\item{series_id}{The ID of the series you want to download}

\item{start_year}{The first (earliest) year for which you want data for the \code{series_id}.
If \code{NULL}, will attempt to pull the earliest available data.}

\item{end_year}{The last (most recent) year for which you want data for the \code{series_id}.
If \code{NULL}, will pull the most recent data.}

\item{api_key}{Your registred BLS API key, or \code{NULL} if you do not have a key.}
}
\value{
A data frame containing the requested data.
}
\description{
Recursively get jobs data from the Bureau of Labor Statistics.
}
\details{
Recursively requests data for the given series. Starts by downloading an \code{increment}
  number of years ending with the user-supplied \code{end_year}, and recursively works backward
  to either the user-supplied \code{start_year} or the first available year, whichever is greater.

  If \code{api_key} is \code{NULL}, \code{get_bls_data()} will set \code{increment = 10},
  working within the 10-year limitation of the BLS API for unregistered users.
}
\seealso{
\code{\link{bls_api}}
}
