% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blscraper_funcs.R
\name{load_bls_data}
\alias{load_bls_data}
\title{load_bls_data}
\usage{
load_bls_data(
  file_name,
  series_id,
  start_year = NULL,
  end_year = NULL,
  api_key = NULL,
  fresh_start = FALSE
)
}
\arguments{
\item{file_name}{character. The path and filename of an R serial data (.RDS) file to use for loading
or saving data on disk.}

\item{series_id}{character. The BLS series ID used to download data.}

\item{start_year}{numeric. The first (earliest) year for which you want data for the \code{series_id}.
If \code{NULL}, will attempt to pull the earliest available data.}

\item{end_year}{numeric. The last (most recent) year for which you want data for the \code{series_id}.
If \code{NULL}, will pull the most recent data.}

\item{api_key}{Your registred BLS API key, or \code{NULL} if you do not have a key.}

\item{fresh_start}{boolean. If true, will delete any existing data file specified by \code{file_name}
before proceeding to download new data from the specified series.}
}
\value{
A data frame
}
\description{
Returns a data frame containing the specified Bureau of Labor Statistics data.
}
\details{
If the file specified by \code{file_name} already exists, the data is loaded
  from disk and returned. If the file does not exist, the series ID is used to download the
  data from the BLS. The data is then saved to the file \code{file_name}, and returned as
  a data frame.
  In addition to the data returned from the BLS's API, a new column \code{date} is added,
  with data class Date, and the data is sorted by date.
  Both \code{file_name} and \code{series_id} should be passed as character strings, e.g.
  \code{file_name = "data/jobs_u.rds", series_id = "CEU0000000001"}.
}
\examples{
  \dontrun{
  # Get all years of seasonally adjusted position reported by employers ("jobs report")
  # and save to data/employer_jobs_s.rds
  employer_jobs_s_df <- load_bls_data(file_name = "data/employer_jobs_s.rds",
                                      series_id = "CES0000000001")

  # Get the last decade of unadjusted employer-reported position
  library(magrittr); library(lubridate)
  my_start_date <- today() \%>\% year()
  employer_decade_df <- load_bls_data(file_name = "data/last_decade_u.rds",
                                      series_id = "CEU0000000001",
                                      start_date = my_start_date)
  }
}
